Handle [Un]LockingTermDeposit
Neither LockingTermDeposit nor UnlockingTermDeposit generate any
output. They represent a move from a Savings Wallet to a Term Wallet
(where a token is held for a fixed period of time in exchange for higher
staking rewards). The only reason for analysing these entries is to have
early warning of potential format changes.

Note that there is a slight issue in that any entries involving a fiat
currency are likely to fail the strict equality test of currency amounts
as, although the fiat currency is recorded exactly, the exchange
equivalent (GBPX in the case of GBP) is reported to 8 decimal digits, so
a simple textual comparison fails. This will be fixed in a subsequent
commit.