Process kraken token purchases
Kraken reports a token purchase as a "spend" entry followed by a
"receive" entry. The "spend" entry specifies the currency used for the
purchase and the amount (plus fee) for the purcahse, with the remaining
details coming from the "receive" entry.

The resulting data is written as a "BUY" record.