Handle Kraken token deposits
Kraken token deposits are now handled correctly and no longer output
twice. The only deposit handling that is not yet complete is handling of
fiat currency deposits, but these do not lead to any output (and extra
handling is just to ensure that the ledger format does not change
unnoticed).

A helper function is used to identify fiat currency. Only fiat
currencies that have been used are recognised as such.

Any new fiat currency usage, which is easonably unlikley,  would lead to
an extra set of outputs identified as a new token, which should be easy
to spot and easily fixable.