Update to handle NEXO transaction format change
Sometime between 6-APR-2022 and the end of APR-2022 the format NEXO uses
in the downloadable transaction history CSV file changed such that
"Exchange" entries now have all the information required to fully record
exactly what happened. Previously a purchase of, for example, 99 units
of FOO using GBP would record FOO and 99 and the USD price but not the
original GBP. Similarly an exchange of 200 FOO-A for 300 FOOB would
record FOO-A, FOO-B and 300, but the 200 was nowhere to be seen. With
the new format these issues are resolved.

This change retains identical output for the older format and tries to
produce enhanced output for the new format.