Ensure that an unknown transaction type is tested for
If a new transaction type is defined, ensure that the code spots this
and reports it loudly in the output stream (where it will deliberately
break the CSV format).

Adjust the main code's error message to better match other warnings and
errors and also make it clear exactly where to look for the error (by
providing the transaction ID). The message now makes it clear that the
problem is that an invalid transaction has been seen.