Update to handle coin transfer
o The format has a new header value "Transaction Hash" (currently always
blank).

o Handle a transfer to another crypto.com user (flagged as "To +" and
then the target phone number

o Handle a reward (flagged as "Pay Reward").

o Suppress the code that forces a BST check - this needs to be handled
in the future