Crypto.com: Add support for exchange of one token for another
The existing code supported an exchange from fiat to a cryptocurrency,
in other words, a purchase. That support has been extended to allow
support for an exchange of one cryptocurrency for another. This is
equivalent to a sell of the first token and a purchase of the second
token. Fortunately crypto.com provides the native currency amount as
part of the record.