Add token price lookup to the Kraken converter
It would save time when building a spreadsheet if the kraken data
included token prices at the time of the entry in the kraken
ledger. Luckily coingecko provides that data so use a library to access
it via its API. The API is rate limited so slow down any lookups to try
to avoid API failures. Cache any data that is looked up to avoid
unnecessary future lookups.