Adjust nexo converter to account for removed "Outstanding Loan" data
Sometime between 2022-04 and 2023-04 the nexo transaction format removed the "Outstanding Loan" column. This column was never used and always read "$0.00". The code no longer expects it to exist and no longer checks it for validity.

Removed a debug fmt.Println that warned ofa format change from 2022-04.

Fixed a trivial typo.