kraken/convert-kraken.go: Remove price lookup code
The coingecko API now wants to be paid for lookups that are longer than a year in the past. I didn't run into this last year but because of the built in caching, perhaps I simply didn't need to do any lookups that far back.

Anyway, on looking for an alternate source, it turns out that it is possible to download full history for any coin they list, as a CSV file. Since I only need a few coins, it's easy to simply download a few files and use those as a source of prices.

For now, however, the simplest thing is to remove the price data so that I can get on with fixing any other changes that kraken have made to the ledger format since I last updated.