kraken/convert-kraken.go: Prepare to e-work all processing
After looking at this for a while, it's clear that the original code cannot be slightly tweaked to process the new format. So I'm making sure that everything that is currently processed no longer handles anything. Anything that currently just logs a fatal is being left untouched. Next I'll gradually work through all the entries that need to be handled and work out what to do with each one.

One quick enhancement is that subtype is only non-blank for earn and transfer records, so enforce that now.