kraken/convert-kraken.go: Handle a subset of "transfer" entries
"transfer" entries never produce any output as they are a record of internal exchange activity. For example the subtype "spottostaking" records a movement out of the spot wallet (headed for the staking wallet). The subtype "stakingfromspot" records the arrival in the staking wallet of something that was previously moved out of the spot wallet. So these should always be paired. The ref-id is of no help in pairing these, so the only way to perform any consistency checking would be to record a "spottostaking" entry and when a "stakingfromspot" entry is seen it should find the first matching entry for the same coin and amount and check that they happened withon some time limit of each other.

Right now all I need is the output data and there are only four "transfer" records and I can see by eye that they are correct. Since they produce no output I'm going to defer introducing the checks and just concentrate on getting good output.

Any unrecognised subtype (i.e. anything other than "spottostaking" or "stakingfromspot" will log a fatal error.