kraken/convert-kraken.go: handle changes in ledger format
Deposits are not listed twice any more. At least not deposits of crypto assets.

The "earn" subtype now includes "deallocation", which can be safely ignored as it is just internal accounting.

Allow further assets to appear in a "spend".

Finally do not complain about being unable to look up the price for USD. Those entries will be ignored anyway. Note that this is USD and not USDC!